% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregranks_summary.R
\name{confint.ivregranks}
\alias{confint.ivregranks}
\title{Title}
\usage{
\method{confint}{ivregranks}(
  object,
  parm,
  level = 0.95,
  component = c("stage2", "stage1"),
  complete = TRUE,
  vcov. = NULL,
  df = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
