% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{createbins}
\alias{createbins}
\title{partition vector into quantile bins}
\usage{
createbins(x, n)
}
\arguments{
\item{x}{vector of values to be partitioned}

\item{n}{number of bins}
}
\value{
vector of the same dimension as \code{x} containing the a bin membership indicator
}
\description{
partition vector into quantile bins
}
