% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{createquartiles}
\alias{createquartiles}
\title{partition vector into quartile bins}
\usage{
createquartiles(x)
}
\arguments{
\item{x}{vector of values to be partitioned}
}
\value{
vector of the same dimension as \code{x} containing the a quartile bin membership indicator
}
\description{
partition vector into quartile bins
}
