% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmeans.R
\name{csdiffmeans}
\alias{csdiffmeans}
\title{Confidence sets for vector of differences}
\usage{
csdiffmeans(x, sd, coverage = 0.95, indices = NA,
  cstype = "symmetric", stepdown = TRUE, R = 1000, seed = NA)
}
\arguments{
\item{x}{vector of estimates}

\item{sd}{vector of standard errors of \code{x}}

\item{coverage}{nominal coverage of the confidence set. Default is 0.95.}

\item{indices}{vector of indices of \code{x} whose differences with all others the confidence set jointly covers. \code{indices=NA} (default) means confidence set jointly covers all possible differences.}

\item{cstype}{type of confidence set (\code{symmetric}, \code{upper}, \code{lower}). Default is \code{symmetric}.}

\item{stepdown}{logical; if \code{TRUE} (default), stepwise procedure is used, otherwise single step procedure is used.}

\item{R}{number of bootstrap replications. Default is 1000.}

\item{seed}{seed for bootstrap random variable draws. If set to \code{NA} (default), then seed is not set.}
}
\value{
L,U lower and upper bounds of the confidence set for ranks indicated in \code{indices}.
}
\description{
Confidence sets for vector of differences
}
