% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csranks-package.R
\docType{package}
\name{csranks-package}
\alias{csranks}
\alias{csranks-package}
\title{csranks: Statistical Tools for Ranks}
\description{
Account for uncertainty when working with ranks. Estimate standard errors consistently in linear regression with ranked variables. Construct confidence sets of various kinds for positions of populations in a ranking based on values of a certain feature and their estimation errors. Theory based on Mogstad, Romano, Shaikh, and Wilhelm (2023)\doi{10.1093/restud/rdad006} and Chetverikov and Wilhelm (2023) \href{https://arxiv.org/abs/2310.15512}{arXiv:2310.15512}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielwilhelm/R-CS-ranks}
  \item \url{https://danielwilhelm.github.io/R-CS-ranks/}
  \item Report bugs at \url{https://github.com/danielwilhelm/R-CS-ranks/issues}
}

}
\author{
\strong{Maintainer}: Daniel Wilhelm \email{d.wilhelm@ucl.ac.uk}

Authors:
\itemize{
  \item Pawel Morgen \email{seriousmorgen@protonmail.com}
}

}
\keyword{internal}
