% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranks.R
\name{csranks_marg}
\alias{csranks_marg}
\title{Marginal confidence sets for ranks}
\usage{
csranks_marg(x, sd, coverage = 0.95, cstype = "two-sided",
  stepdown = TRUE, R = 1000, indices = NA, na.rm = FALSE,
  seed = NA)
}
\arguments{
\item{x}{vector of estimates}

\item{sd}{vector of standard errors of \code{x}}

\item{coverage}{nominal coverage of the confidence set. Default is 0.95.}

\item{cstype}{type of confidence set (\code{two-sided}, \code{upper}, \code{lower}). Default is \code{two-sided}.}

\item{stepdown}{logical; if \code{TRUE} (default), stepwise procedure is used, otherwise single step procedure is used.}

\item{R}{number of bootstrap replications. Default is 1000.}

\item{indices}{vector of indices of \code{x} for whose ranks the confidence sets are computed. \code{indices=NA} (default) means computation for all ranks.}

\item{na.rm}{logical; if \code{TRUE}, then \code{NA}'s are removed from \code{x} and \code{sd} (if any).}

\item{seed}{seed for bootstrap random variable draws. If set to \code{NA} (default), then seed is not set.}
}
\value{
L, U lower and upper bounds of the confidence set for ranks indicated in \code{indices}.
}
\description{
This function is called by \code{\link{csranks}} when \code{simul=FALSE}.
}
\section{Details}{

Implentation of the simultaneous confidence sets proposed in Mogstad, Romano, Shaikh, and Wilhelm ("Inference for Ranks with Applications to Mobility across Neighborhoods and Academic Achievements across Countries", \href{https://www.ucl.ac.uk/~uctpdwi/papers/cwp1020.pdf}{CeMMAP Working Paper CWP10/20}).
}

