% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranks_multinom.R
\name{csranks_multinom_simul}
\alias{csranks_multinom_simul}
\title{Simultaneous Confidence sets for ranks based on multinomial data}
\usage{
csranks_multinom_simul(x, coverage = 0.95, cstype = "two-sided",
  multcorr = "Bonferroni", indices = NA, na.rm = FALSE)
}
\arguments{
\item{x}{vector of counts of successes for each category}

\item{coverage}{nominal coverage of the confidence set. Default is 0.95.}

\item{cstype}{type of confidence set (\code{two-sided}, \code{upper}, \code{lower}). Default is \code{two-sided}.}

\item{multcorr}{multiplicity correction to be used: \code{Bonferroni} (default) or \code{Holm}.}

\item{indices}{vector of indices of \code{x} for whose ranks the confidence sets are computed. \code{indices=NA} (default) means computation for all ranks.}

\item{na.rm}{logical; if \code{TRUE}, then \code{NA}'s are removed from \code{x} and \code{sd} (if any).}
}
\value{
L, U lower and upper bounds of the confidence set for ranks indicated in \code{indices}.
}
\description{
This function is called by \code{\link{csranks_multinom}} when \code{simul=TRUE}.
}
