% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{frank}
\alias{frank}
\title{Compute fractional ranks}
\usage{
frank(x, omega = 0, increasing = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{vector of estimates}

\item{na.rm}{logical; if \code{TRUE}, then \code{NA}'s are removed from \code{x} and \code{sd} (if any).}
}
\value{
vector of the same dimension as \code{x} containing the ranks
}
\description{
Compute fractional ranks
}
\examples{
frank(c(4,4,4,3,1,10,7,7))
}
