% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parent_child_income}
\alias{parent_child_income}
\title{Income of parents and children}
\format{
A data frame with 3894 rows and 4 variables:
\describe{
\item{c_faminc}{Family income of a child}
\item{p_faminc}{Family income of parent}
\item{gender}{Gender}
\item{race}{Race: hisp (Hispanic), black or neither}
}
}
\usage{
parent_child_income
}
\description{
An artificial dataset containing income of children and their parents together
with some information about them.
}
\keyword{datasets}
