% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa2018}
\alias{pisa2018}
\alias{pisa2022}
\title{Cross-country comparison of students' achievement}
\format{
\describe{
\item{jurisdiction}{country, from which data was collected}
\item{math_score}{average score in math}
\item{math_se}{standard error for the average score in math}
\item{reading_score}{average score in reading}
\item{reading_se}{standard error for the average score in reading}
\item{science_score}{average score in science}
\item{science_se}{standard error for the average score in science}
}

An object of class \code{data.frame} with 38 rows and 7 columns.
}
\source{
\url{https://www.oecd.org/en/about/programmes/pisa/pisa-data.html}
}
\usage{
pisa2018

pisa2022
}
\description{
Datasets containing average scores on math, reading, and science
together with standard errors for all OECD countries. These are
from the 2018 and 2022 editions of
Program for International Student Assessment (PISA)
study by the Organization for Economic Cooperation and Development (OECD).
The average scores are over all 15-year-old students in the study.
}
\keyword{datasets}
