% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrreg.R
\name{simple_lmranks}
\alias{simple_lmranks}
\title{Linear model for ranks}
\usage{
simple_lmranks(Y, X, W = NULL, omega = 0, increasing = FALSE, na.rm = FALSE)
}
\arguments{
\item{Y}{Numeric vector. Based on it, ranks for response will be constructed
and used as final response in the linear model.}

\item{X}{Numeric vector. Based on it, ranks for this covariate will be constructed
and used as final covariate in the linear model.}

\item{W}{Matrix. Other covariates to include in the model. Should not include an intercept column.}

\item{omega}{numeric; numeric value in [0,1], each corresponding to a different definition of the rank; default is \code{0}. See \code{\link{frank}} for details.}

\item{increasing}{logical; if \code{TRUE}, then large elements in \code{X} and \code{Y} receive a large rank. Otherwise, large elements receive small ranks.}

\item{na.rm}{logical; if \code{TRUE}, then \code{NA}'s are removed from \code{X} and \code{Y} (if any).}
}
\value{
A list with two items: 
\itemize{
\item{\code{rhohat} - the linear coefficient for rank of X.}
\item{\code{se} - Estimated standard error of the coefficient rhohat.}
}
}
\description{
Fit a linear model with a single rank response and
a single rank covariate (and possibly other usual covariates).
}
\note{
TODO: make it more similar to lm.
}
