% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregranks_summary.R
\name{summary.ivregranks}
\alias{summary.ivregranks}
\title{Title}
\usage{
\method{summary}{ivregranks}(object, vcov. = NULL, df = NULL, diagnostics = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
