% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrreg_methods.R
\name{summary.lmranks}
\alias{summary.lmranks}
\title{Summarizing fits of linear models for ranks}
\usage{
\method{summary}{lmranks}(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\description{
Summary method for class "\code{lmranks}". It returns theoretically valid standard
errors
}
