% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregranks_summary.R
\name{vcov.ivregranks}
\alias{vcov.ivregranks}
\title{Title}
\usage{
\method{vcov}{ivregranks}(object, component = c("stage2", "stage1"), complete = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
