% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunctions.R
\name{xrank}
\alias{xrank}
\title{Compute ranks}
\usage{
xrank(x, best = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{vector of values to be ranked}

\item{best}{logical; if \code{TRUE} (default), the rank of the j-th element is defined as the number of other elements strictly larger than the j-th. Otherwise, the rank is defined as the number of other elements strictly smaller than the j-th.}

\item{na.rm}{logical; if \code{TRUE}, then \code{NA}'s are removed from \code{x} (if any).}
}
\value{
vector of the same dimension as \code{x} containing the ranks
}
\description{
Compute ranks
}
