% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranks.R
\name{csranks}
\alias{csranks}
\alias{csranks_simul}
\alias{csranks_marg}
\title{Confidence sets for ranks}
\usage{
csranks(
  x,
  sd,
  coverage = 0.95,
  cstype = "two-sided",
  stepdown = TRUE,
  R = 1000,
  simul = TRUE,
  indices = NA,
  na.rm = FALSE,
  seed = NA
)

csranks_simul(
  x,
  sd,
  coverage = 0.95,
  cstype = "two-sided",
  stepdown = TRUE,
  R = 1000,
  indices = NA,
  na.rm = FALSE,
  seed = NA
)

csranks_marg(
  x,
  sd,
  coverage = 0.95,
  cstype = "two-sided",
  stepdown = TRUE,
  R = 1000,
  indices = NA,
  na.rm = FALSE,
  seed = NA
)
}
\arguments{
\item{x}{vector of estimates.}

\item{sd}{vector of standard errors of \code{x} OR a covariance matrix of vector \code{x}.}

\item{coverage}{nominal coverage of the confidence set. Default is 0.95.}

\item{cstype}{type of confidence set (\code{two-sided}, \code{upper}, \code{lower}). Default is \code{two-sided}.}

\item{stepdown}{logical; if \code{TRUE} (default), stepwise procedure is used, otherwise single step procedure is used. See Details section for more.}

\item{R}{number of bootstrap replications. Default is 1000.}

\item{simul}{logical; if \code{TRUE} (default), then simultaneous confidence sets are computed, which jointly cover all populations indicated by \code{indices}.
Otherwise, for each population indicated in \code{indices} a marginal confidence set is computed.}

\item{indices}{vector of indices of \code{x} for whose ranks the confidence sets are computed. \code{indices=NA} (default) means computation for all ranks.}

\item{na.rm}{logical; if \code{TRUE}, then \code{NA}'s are removed from \code{x} and \code{sd} (if any).}

\item{seed}{seed for bootstrap random variable draws. If set to \code{NA} (default), then seed is not set.}
}
\value{
A list with two items, `L` and `U` - lower and upper bounds of the confidence set for ranks indicated in \code{indices}.
}
\description{
Given estimates and their standard errors (or covariance matrix) of a certain feature for a set of populations,
calculate confidence sets for the ranks of populations,
where populations are ranked by the feature values.
}
\section{Functions}{
\itemize{
\item \code{csranks_simul()}: Simultaneous confidence sets for ranks

This function is called by \code{csranks} when \code{simul=TRUE}.

\item \code{csranks_marg()}: Marginal confidence sets for ranks

This function is called by \code{csranks} when \code{simul=FALSE}.

}}
\section{Details}{

The command implements the procedure for construction of confidence sets for ranks described in the referenced paper below.
Generally, it consists of verification of a large set of hypotheses. After rejection of certain set
of hypotheses, one can terminate the procedure or keep verifying a smaller set of hypotheses that
were not rejected so far. The former corresponds to \code{stepdown=FALSE}; the latter to \code{stepdown=TRUE}.

From a practical point of view, \code{stepdown=TRUE} takes more time, but usually results
in tighter (better) confidence sets.

Parametric bootstrap used to calculate distribution for confidence sets based on the multivariate normal distribution.
}

\examples{
x <- seq(1, 3, length = 10)
sd <- rep(0.2, 10)
csranks(x, sd)
csranks_simul(x, sd)
csranks_marg(x, sd)
}
\references{
Mogstad, Romano, Shaikh, and Wilhelm.
"Inference for Ranks with Applications to Mobility across Neighborhoods and Academic Achievements across Countries"
\href{https://www.ucl.ac.uk/~uctpdwi/papers/cwp1020.pdf}{CeMMAP Working Paper CWP10/20}
}
