% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregranks.R
\name{ivregranks}
\alias{ivregranks}
\title{Title}
\usage{
ivregranks(
  formula,
  instruments,
  data,
  subset,
  na.action,
  weights,
  offset,
  contrasts = NULL,
  model = TRUE,
  y = TRUE,
  x = FALSE,
  method = c("OLS", "M", "MM"),
  omega = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
