% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregranks.R, R/ivregranks_summary.R
\name{ivregranks}
\alias{ivregranks}
\alias{summary.ivregranks}
\alias{print.summary.ivregranks}
\alias{confint.ivregranks}
\alias{vcov.ivregranks}
\title{Instrumental-Variable Regression by 2SLS, 2SM, or 2SMM Estimation Involving
Ranks}
\usage{
ivregranks(
  formula,
  instruments,
  data,
  subset,
  na.action,
  weights,
  offset,
  contrasts = NULL,
  model = TRUE,
  y = TRUE,
  x = FALSE,
  method = c("OLS", "M", "MM"),
  omega = 1,
  ...
)

\method{summary}{ivregranks}(object, vcov. = NULL, df = NULL, diagnostics = NULL, ...)

\method{print}{summary.ivregranks}(x, ...)

\method{confint}{ivregranks}(
  object,
  parm,
  level = 0.95,
  component = c("stage2", "stage1"),
  complete = TRUE,
  vcov. = NULL,
  df = NULL,
  ...
)

\method{vcov}{ivregranks}(object, component = c("stage2", "stage1"), complete = TRUE, ...)
}
\arguments{
\item{formula, instruments}{formula specification(s) of the regression
relationship and the instruments. Either \code{instruments} is missing and
\code{formula} has three parts as in \code{r(y) ~ x1 + r(x2) | r(z1) + z2 +
z3} (recommended) or \code{formula} is \code{r(y) ~ x1 + r(x2)} and
\code{instruments} is a one-sided formula \code{~ z1 + z2 + z3} (only for
backward compatibility).}

\item{data}{an optional data frame containing the variables in the model.
By default the variables are taken from the environment of the
\code{formula}.}

\item{subset}{currently not supported.}

\item{na.action}{currently not supported.}

\item{weights}{currenty not supported.}

\item{offset}{an optional offset that can be used to specify an a priori
known component to be included during fitting.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of
\code{\link[stats:model.matrix]{model.matrix.default}}.}

\item{model, x, y}{logicals.  If \code{TRUE} the corresponding components of
the fit (the model frame, the model matrices, the response) are returned.
These components are necessary for computing regression diagnostics.}

\item{method}{the method used to fit the stage 1 and 2 regression:
\code{"OLS"} for traditional 2SLS regression (the default),
\code{"M"} for M-estimation, or \code{"MM"} for MM-estimation, with the
latter two robust-regression methods implemented via the
\code{\link[MASS]{rlm}} function in the \pkg{MASS} package.}

\item{\dots}{further arguments passed to \code{\link[ivreg]{ivreg.fit}}.}

\item{object}{An object of class \code{"ivregranks"}.}

\item{vcov.}{Optionally either a coefficient covariance matrix or a function to compute such a covariance
matrix from fitted \code{ivreg} model objects. If \code{NULL} (the default) the standard covariance matrix
(based on the information matrix) is used. Alternatively, covariance matrices (e.g., clustered and/or
heteroscedasticity-consistent) can be plugged in to adjust Wald tests or confidence intervals etc.
In \code{summary}, if \code{diagnostics = TRUE}, \code{vcov.} must be a function (not a matrix) because
the alternative covariances are also needed for certain auxiliary models in the diagnostic tests.
If \code{vcov.} is a function, the \code{...} argument can be used to pass on further arguments to
this function.}

\item{df}{For \code{summary}, optional residual degrees of freedom to use in computing model summary.}

\item{diagnostics}{Report 2SLS "diagnostic" tests in model summary (default is \code{TRUE}). 
These tests are not to be confused with the \emph{regression diagnostics} provided elsewhere in the \pkg{ivreg}
package: see \code{\link[ivreg]{ivregDiagnostics}}.}

\item{parm}{parameters for which confidence intervals are to be computed; a vector or numbers or names; the default is all parameters.}

\item{level}{confidence level; the default is \code{0.95}.}

\item{component}{Character indicating \code{"stage2"} or \code{"stage1"}.}

\item{complete}{logical indicating if the full variance-covariance matrix
should be returned also in case of an over-determined system where
some coefficients are undefined and \code{coef(.)} contains NAs
correspondingly. When \code{complete = TRUE}, \code{vcov()} is compatible
with \code{coef()} also in this singular case.}
}
\value{
\code{ivregranks} returns an object of class \code{"ivregranks"} that
inherits as much as possible from class \code{"ivreg"}, with the following
additional components:
\item{rank_term_indices}{an integer vector with indices of entries of
\code{terms.labels} attribute of \code{terms(formula) for the outcome
equation which correspond to ranked regressors.}
\item{ranked_instrument_indices}{an integer vector with indices of entries
of the ranked instrumental variables.}
\item{ranked_response}{a logical entry.}
\item{omega}{an entry corresponding to the \code{omega} argument.}
}
\description{
Fit instrumental-variable regression involving ranks by two-stage least
squares (2SLS). This is equivalent to direct instrumental-variables
estimation when the number of instruments is equal to the number of
regressors. Alternative robust-regression estimators are also provided,
based on M-estimation (2SM) and MM-estimation (2SMM).
}
\details{
\code{ivregranks} is the high-level interface to the work-horse function
A set of standard methods (including \code{print}, \code{summary},
\code{vcov}, \code{anova}, \code{predict}, \code{residuals}, \code{terms},
\code{model.matrix}, \code{bread}, \code{estfun}) is available and described
in \code{\link{ivregranks_summary}}.

Regressors and instruments for \code{ivregranks} are most easily specified
in a formula with two parts on the right-hand side, e.g.,
\code{r(y) ~ x1 + r(x2) | r(z1) + z2 + z3}, where \code{x1} and \code{r(x2)}
are the explanatory variables and \code{r(z1)}, \code{z2}, and \code{z3} are
the instrumental variables. Note that exogenous regressors have to be
included as instruments for themselves.

For example, if there is
one exogenous regressor \code{ex} and one endogenous regressor \code{r(en)}
with instrument \code{r(in)}, the appropriate formula would be \code{r(y) ~
r(en) + ex | r(in) + ex}. Alternatively, a formula with three parts on the
right-hand side can also be used: \code{r(y) ~ ex | r(en) | r(in)}.
The latter is typically more convenient, if there is a large number of
exogenous regressors.

Moreover, two further equivalent specification strategies are possible that
are typically less convenient compared to the strategies above. One option
is to use an update formula with a \code{.} in the second part of the formula
is used: \code{r(y) ~ r(en) + ex | . - r(en) + r(in)}. Another option is to
use a separate formula for the instruments (only for backward compatibility
with earlier versions):
\code{formula = r(y) ~ r(en) + ex, instruments = ~ r(in) + ex}.

Internally, all specifications are converted to the version with two parts
on the right-hand side.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary(ivregranks)}: Summary and Inference Methods for \code{"ivregranks"}
Objects

\item \code{vcov(ivregranks)}: Calculate Variance-Covariance Matrix for a Fitted
\code{ivregranks} object

Returns the variance-covariance matrix of the regression coefficients
(main parameters) of a fitted \code{ivregranks} object. Its result is
theoretically valid and asymptotically consistent, in contrast to naively
running \code{vcov(ivreg(...))}.

}}
\seealso{
\code{\link[ivreg]{ivreg.fit}}, \code{\link[csranks]{lmranks}}
}
\keyword{.}
\keyword{Generic}
\keyword{\code{\link[stats]{coef}},}
\keyword{\code{\link[stats]{effects}},}
\keyword{\code{\link[stats]{fitted}},}
\keyword{\code{\link[stats]{model.frame}},}
\keyword{\code{\link[stats]{model.matrix}},}
\keyword{\code{\link[stats]{residuals}},}
\keyword{\code{\link[stats]{update}}}
\keyword{functions}
\keyword{regression}
