% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmranks.R
\name{lmranks}
\alias{lmranks}
\title{Linear model for ranks}
\usage{
lmranks(
  formula,
  data,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  constrasts = NULL,
  offset = offset,
  omega = 0,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{formula}{An object of class "\code{\link{formula}}": a symbolic description
of the model to be fitted. Exactly like the formula for linear model except that
rank terms, \code{r()}, can be added to specify that the linear regressor depends on ranks of regressors
and to specify a rank response. See Details.}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{lm} is called.}

\item{subset}{an optional vector specifying a subset of observations to be used 
in the fitting process. The ranks will be calculated using full data. 
(See additional details about how this argument interacts with data-dependent 
bases in the ‘Details’ section of the \code{\link{model.frame}} documentation.)}

\item{weights}{currently not supported.}

\item{na.action}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is set by
    the \code{na.action} setting of \code{\link{options}}, and is
    \code{\link[stats]{na.fail}} if that is unset.  The \sQuote{factory-fresh}
    default is \code{\link[stats]{na.omit}}.  Another possible value is
    \code{NULL}, no action.  Value \code{\link[stats]{na.exclude}} can be useful.}

\item{method}{the method to be used; for fitting, currently only
    \code{method = "qr"} is supported; \code{method = "model.frame"} returns
    the model frame (the same as with \code{model = TRUE}, see below).}

\item{singular.ok}{logical. If \code{FALSE} (the default in S but
    not in \R) a singular fit is an error.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector or matrix of extents
    matching those of the response.  One or more \code{\link[stats]{offset}} terms can be
    included in the formula instead or as well, and if more than one are
    specified their sum is used.  See \code{\link[stats]{model.offset}}.}

\item{omega}{as in \code{\link{frank}}.}

\item{na.rm}{If \code{FALSE}, raises an error is any \code{NA} values in ranked regressors or response
are encountered. If \code{TRUE}, ranks for non-\code{NA} entries are calculated by ignoring \code{NA} values.
For \code{NA} values, \code{NA} ranks are returned and handled later by \code{na.action}.}

\item{...}{For \code{lm()}: additional arguments to be passed to the low level
    regression fitting functions (see below).}
}
\value{
An object of class \code{lmranks}, inheriting (as well as possible) from class \code{lm}.
See the \code{\link{lm}} documentation for more.

Additionally, it has an \code{omega} entry, corresponding to \code{omega} argument,
and a \code{rank_terms_indices} - an integer vector with indices of entries of \code{terms.labels} attribute
of \code{terms(formula)}, which correspond to ranked regressors.

A number of methods defined for \code{lm} does not yield theoretically correct 
results when applied to \code{lmranks} objects; errors or warnings are raised consciously.
Also, the \code{df.residual} component is set to NA, since the notion of effects of freedom
for the rank models is not theoretically established.
}
\description{
Fit a linear model with a single rank response and
a single rank regressor (and possibly other usual regressors).
}
\details{
This function is useful in case when relationship not between variables themselves, but their rank
(or, put differently, their ECDF value) is of interest. The variables to be ranked
(both dependent and independent) can be marked with \code{r()}. 

The \code{r()} is a private alias for \code{\link{frank}} and can accept 
\code{increasing} argument. However, \code{omega} argument must be specified globally 
(as a specification of rank definition) in the call to \code{lmranks}.

As a consequence of the order, in which model.frame applies operations, \code{subset} 
and \code{na.action} are applied after evaluation of \code{r()}. This means, that
1) the ranks will be calculated using full data. In order to calculate them on subsetted data,
one may subset the data outside of lm and pass it simply as new \code{data} argument.
2) \code{na.action} will not handle NA values in ranked regressors. This means,
that they have to be handled separately by the user.

Currently, only models with single rank response, single rank regressor and
(possibly) other usual regressors is available.
}
\section{Warning}{

Wrapping \code{r()} with other functions (like \code{log(r(x))}) will not 
recognize correctly the mark (because it will not be caught in \code{terms(formula, specials = "r")}).
The ranks will be calculated correctly, but their transformation will be treated later in \code{lm} as a regular
regressor. This means, that the corresponding regression coefficient will be calculated correctly,
but the standard errors, statistics etc. will not. 

\code{r}, \code{.r_predict} and \code{.r_cache} are special expressions, used
internally to interpret \code{r} mark correctly. Do not use them in \code{formula}.
}

\examples{
Y <- c(3,1,2,4,5)
y_frank <- c(0.6, 1.0, 0.8, 0.4, 0.2)
X <- 1:5
omega <- 0.5
x_frank <- c(1.0, 0.8, 0.6, 0.4, 0.2)
W <- matrix(y_frank * 0.1 + 5 + rnorm(5, sd = 0.1), ncol = 1)

lmranks(r(Y) ~ r(X) + W)
# naive version with same regression coefficients, but incorrect 
# standard errors and statistics:
lm(y_frank ~ x_frank + W)

data(mtcars)
lmranks(r(mpg) ~ r(hp) + ., data = mtcars)

}
\seealso{
\code{\link{lm}} for details about other arguments; \code{\link{frank}}.
}
