% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrreg.R
\name{lmranks}
\alias{lmranks}
\title{Linear model for ranks}
\usage{
lmranks(
  formula,
  data,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  constrasts = NULL,
  offset = offset,
  omega = 0,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{formula}{An object of class "\code{\link{formula}}": a symbolic description
of the model to be fitted. Exactly like the formula for linear model except that
rank terms, \code{r()}, can be added to specify that the linear predictor depends on ranks of predictors
and to specify a rank response. See Details.}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{lm} is called.}

\item{subset}{an optional vector specifying a subset of observations
    to be used in the fitting process.  (See additional details about how
    this argument interacts with data-dependent bases in the
    \sQuote{Details} section of the \code{\link[stats]{model.frame}}
    documentation.)}

\item{weights}{an optional vector of weights to be used in the fitting
    process.  Should be \code{NULL} or a numeric vector.
    If non-NULL, weighted least squares is used with weights
    \code{weights} (that is, minimizing \code{sum(w*e^2)}); otherwise
    ordinary least squares is used.  See also \sQuote{Details},}

\item{na.action}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is set by
    the \code{na.action} setting of \code{\link{options}}, and is
    \code{\link[stats]{na.fail}} if that is unset.  The \sQuote{factory-fresh}
    default is \code{\link[stats]{na.omit}}.  Another possible value is
    \code{NULL}, no action.  Value \code{\link[stats]{na.exclude}} can be useful.}

\item{method}{the method to be used; for fitting, currently only
    \code{method = "qr"} is supported; \code{method = "model.frame"} returns
    the model frame (the same as with \code{model = TRUE}, see below).}

\item{singular.ok}{logical. If \code{FALSE} (the default in S but
    not in \R) a singular fit is an error.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector or matrix of extents
    matching those of the response.  One or more \code{\link[stats]{offset}} terms can be
    included in the formula instead or as well, and if more than one are
    specified their sum is used.  See \code{\link[stats]{model.offset}}.}

\item{omega}{as in \code{\link{frank}}.}

\item{...}{For \code{lm()}: additional arguments to be passed to the low level
    regression fitting functions (see below).}
}
\value{
An object of class \code{lmranks}, inheriting (as well as possible) from class \code{lm}.
See the \code{\link{lm}} documentation for more.

Additionally, it has an \code{omega} entry, corresponding to \code{omega} argument,
and a \code{rank_terms_indices} - an integer vector with indices of entries of \code{terms.labels} attribute
of \code{terms(formula)}, which correspond to ranked covariates.

A number of methods defined for \code{lm} does not yield theoretically correct 
results when applied to \code{lmranks} objects; errors or warnings are raised consciously.
Also, the \code{df.residual} component is set to NA, since the notion of effects of freedom
for the rank models is not theoretically developed.
}
\description{
Fit a linear model with a single rank response and
a single rank covariate (and possibly other usual covariates).
}
\details{
This function is useful in case when relationship not between variables themselves, but their rank
(or, put differently, their ECDF value) is of interest. The variables to be ranked
(both dependent and independent) can be marked with \code{r()}. 

The \code{r()} is a private alias for \code{\link{frank}} and can accept 
\code{increasing} argument. However, \code{omega} argument must be specified globally 
(as a specification of rank definition) in the call to \code{lmranks}.

Currently, only models with single rank response, single rank covariate and
(possibly) other usual covariates is available.
}
\examples{
Y <- c(3,1,2,4,5)
y_frank <- c(0.6, 1.0, 0.8, 0.4, 0.2)
X <- 1:5
omega <- 0.5
x_frank <- c(1.0, 0.8, 0.6, 0.4, 0.2)
W <- matrix(y_frank * 0.1 + 5 + rnorm(5, sd = 0.1), ncol = 1)

lmranks(r(Y) ~ r(X) + W)
# equivalent:
lm(y_frank ~ x_frank + W)

data(mtcars)
lmranks(r(mpg) ~ r(hp) + ., data = mtcars)

}
\seealso{
\code{\link{lm}} For details about other arguments; \code{\link{frank}}.
}
