% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotranking.R
\name{plotranking}
\alias{plotranking}
\title{Plot ranking with confidence sets}
\usage{
plotranking(
  ranks,
  L,
  U,
  popnames = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  colorbins = 1,
  horizontal = TRUE
)
}
\arguments{
\item{ranks}{vector of ranks}

\item{L}{vector of lower bounds of confidence sets for the ranks}

\item{U}{vector of lower bounds of confidence sets for the ranks}

\item{popnames}{vector containing names of the populations whose ranks are in \code{ranks}. If \code{popnames=NULL} (default), then populations are automatically numbered.}

\item{title}{character string containing the main title of the graph. \code{title=NULL} (default) means no title.}

\item{subtitle}{character string containing the subtitle of the graph. \code{subtitle=NULL} (default) means no subtitle.}

\item{caption}{character string containing the caption of the graph. \code{caption=NULL} (default) means no caption.}

\item{colorbins}{integer indicating the number of quantile bins into which populations are grouped and color-coded. Value has to lie between 1 (default) and the number of populations.}

\item{horizontal}{logical. Should be the bars displayed horizontally, or vertically?}
}
\value{
A ggplot plot displaying confidence sets.
}
\description{
Plot ranking with confidence sets
}
\examples{
x <- seq(1,3,length=10)
sd <- rep(0.2,10)
ranks <- xrank(x)
CS <- csranks(x, sd)
grid::current.viewport()
plotranking(ranks, CS$L, CS$U)
}
