% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotranking.R
\name{plotranking}
\alias{plotranking}
\title{Plot ranking with confidence sets}
\usage{
plotranking(
  ranks,
  L,
  U,
  popnames = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  colorbins = 1,
  horizontal = TRUE
)
}
\arguments{
\item{ranks}{vector of ranks}

\item{L}{vector of lower bounds of confidence sets for the ranks}

\item{U}{vector of lower bounds of confidence sets for the ranks}

\item{popnames}{vector containing names of the populations whose ranks are in \code{ranks}. If \code{popnames=NULL} (default), then populations are automatically numbered.}

\item{title}{character string containing the main title of the graph. \code{title=NULL} (default) means no title.}

\item{subtitle}{character string containing the subtitle of the graph. \code{subtitle=NULL} (default) means no subtitle.}

\item{caption}{character string containing the caption of the graph. \code{caption=NULL} (default) means no caption.}

\item{colorbins}{integer indicating the number of quantile bins into which populations are grouped and color-coded. Value has to lie between 1 (default) and the number of populations.}

\item{horizontal}{logical. Should be the bars displayed horizontally, or vertically?}
}
\value{
A ggplot plot displaying confidence sets.
}
\description{
Plot ranking with confidence sets
}
\examples{
x <- seq(1, 3, length = 10)
V <- diag(rep(0.04, 10))
ranks <- xrank(x)
CS <- csranks(x, V)
grid::current.viewport()
plotranking(ranks, CS$L, CS$U)

# plotranking returns a ggplot object. It can be customized further:
library(ggplot2)
pl <- plotranking(ranks, CS$L, CS$U)
pl + xlab("position in ranking") + ylab("population label") + theme_gray()

# horizontal = FALSE uses ggplot2::coord_flip underneath. The x and y axes swap places.
pl <- plotranking(ranks, CS$L, CS$U, horizontal = FALSE)
pl + xlab("position in ranking") + # Note, that xlab refers to vertical axis now
  ylab("population label") + theme_gray()

}
