% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmranks.R
\name{simple_lmranks}
\alias{simple_lmranks}
\alias{simple_lmranks_rho_se}
\title{Linear model for ranks}
\usage{
simple_lmranks(Y, X, W = NULL, omega = 0, increasing = FALSE, na.rm = FALSE)

simple_lmranks_rho_se(
  Y,
  X,
  W = NULL,
  omega = 0,
  increasing = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{Y}{Numeric vector. Based on it, ranks for response will be constructed
and used as final response in the linear model.}

\item{X}{Numeric vector. Based on it, ranks for this regressor will be constructed
and used as final regressor in the linear model.}

\item{W}{Matrix. Other regressors to include in the model. Should not include an intercept column.}

\item{omega}{numeric; numeric value in [0,1], each corresponding to a different definition of the rank; default is \code{0}. See \code{\link{frank}} for details.}

\item{increasing}{logical; if \code{TRUE}, then large elements in \code{X} and \code{Y} receive a large rank. Otherwise, large elements receive small ranks.}

\item{na.rm}{logical; if \code{TRUE}, then \code{NA}'s are removed from \code{X} and \code{Y} (if any).}
}
\value{
A list with two items: 
\itemize{
\item{\code{rhohat} - the linear coefficient for rank of X.}
\item{\code{se} - Estimated standard error of the coefficient rhohat.}
}
}
\description{
Fit a linear model with a single rank response and
a single rank regressor (and possibly other usual regressors).
}
\section{Functions}{
\itemize{
\item \code{simple_lmranks_rho_se()}: Calculate the standard error of rank regression coefficient

}}
\examples{
Y <- c(3,1,2,4,5)
y_frank <- c(0.6, 1.0, 0.8, 0.4, 0.2)
X <- 1:5
omega <- 0.5
x_frank <- c(1.0, 0.8, 0.6, 0.4, 0.2)
W <- matrix(y_frank * 0.1 + 5 + rnorm(5, sd = 0.1), ncol = 1)

simple_lmranks(Y, X, W)

}
